Add selection and inspector functionality to MyGUI
Added a pointer `selectedGameObject` to the `MyGUI` class to track the currently selected game object. Modified the `ImGui::Selectable` call in the `MyGUI::render` method to update `selectedGameObject` when an item is selected. Added a new ImGui window named "Inspector" to the `MyGUI::render` method, which displays details of the selected game object. The "Inspector" window shows the name and position of the selected game object and allows the user to modify its position. If no game object is selected, the "Inspector" window displays a message indicating that no game object is selected.